word_generator/requirements.txt
functions-framework==3.4.0
google-cloud-firestore==2.11.1
google-generativeai==0.3.1